passage de fonctions en const et fix d un leak sur gnl (attention a la norme)

+++ ft_putfloat.c
@@ -6,7 +6,7 @@
 /*   By: snicolet <snicolet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/02/26 23:01:09 by snicolet          #+#    #+#             */
-/*   Updated: 2016/02/26 23:20:00 by snicolet         ###   ########.fr       */
+/*   Updated: 2016/03/24 12:52:30 by snicolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@ void	ft_putfloat(float nb)
 {
 	char	buffer[32];
 
-	write(1, buffer, (size_t)ft_ftobuff(buffer, nb, 10, "0123456789"));
+	write(1, buffer, (size_t)ft_ftobuff(buffer, (double)nb, 10, "0123456789"));
 }
